+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Add a dict of URIs to replace in a RDF graph at harvest time [#2884](https://github.com/opendatateam/udata/pull/2884)
 - Fix duplicate recipients in new comments mail [#2886](https://github.com/opendatateam/udata/pull/2886)
 - Add type to resource csv adapter [#2888](https://github.com/opendatateam/udata/pull/2888)
+- Return 400 instead of 500 in case of not ObjectID arg in API [#2889](https://github.com/opendatateam/udata/pull/2889)
 
 ## 6.1.6 (2023-07-19)